+++ tests/src/rules/no-duplicates.js
@@ -711,6 +711,25 @@ context('TypeScript', function () {
             },
           ],
         }),
+        // #2834 Detect duplicates across type and regular imports
+        test({
+          code: "import {AValue} from './foo'; import type {AType} from './foo'",
+          ...parserConfig,
+          options: [{ 'prefer-inline': true }],
+          output: `import {AValue,type AType} from './foo'; `,
+          errors: [
+            {
+              line: 1,
+              column: 22,
+              message: "'./foo' imported multiple times.",
+            },
+            {
+              line: 1,
+              column: 56,
+              message: "'./foo' imported multiple times.",
+            },
+          ],
+        }),
       ]);
 
       ruleTester.run('no-duplicates', rule, {